Misc bug fixes

+++ delete-question.php
@@ -1,17 +1,20 @@
 <?php
     require_once(dirname(__FILE__)."/init.php");
+    $questionID = $_GET["id"];
 
-    if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+    $query = 'SELECT Question, Answer FROM Questions WHERE QuestionID = ?';
+    $stmt = $pdo->prepare($query);
+    $stmt->execute([$questionID]);
+    $question = $stmt->fetch();
+    if ($question == NULL) {
+        die("invalid question id");
+    }
+    
+    if ($_SERVER['REQUEST_METHOD'] === 'POST' && $questionID == $_POST["question-id"]) {
         $query = 'DELETE FROM Questions WHERE QuestionID = ?';
         $stmt = $pdo->prepare($query);
-        $stmt->execute([$_GET["id"]]);
+        $stmt->execute([$questionID]);
         header("Location: view-questions.php");
-    }
-    else {
-        $query = 'SELECT Question, Answer FROM Questions WHERE QuestionID = ?';
-        $stmt = $pdo->prepare($query);
-        $stmt->execute([$_GET["id"]]);
-        $question = $stmt->fetch();
     }
 
 ?>
@@ -23,10 +26,8 @@
 <div id="delete-user">
     <p> Are you sure you want to delete the question '<?= $question["Question"] ?>' with answer <?= $question["Answer"] ?>? </p>
     <form method="post">
-        <input type="hidden" name="question-id" value="<?= $_GET['id'] ?>"/>
-        <p>
-            <input type="submit" value="Delete Question"/>
-        </p>
+        <input type="hidden" name="question-id" value="<?= $questionID ?>"/>
+        <button class="btn waves-effect waves-light submit red white-text" type="submit" name="action">Delete Question</button>
     </form>
 </div>
